Separate email sent page from sign up
This makes the "browser asks you to save password" flow more consistent with other sites.
